Add app_path function for cross-platform pathing
PyInstaller mostly works, but we need to make it possible
for the "app-ified" game to locate the assets relative to
itself - PyInstaller adds special `sys` attributes to help.

(And we need to express that path using the appropriate
path separator characters for the OS we're running on!)

This seems to fix it for MacOS, hopefully also for Windows?
(I'll confirm later tonight)

Not sure if there's a better name for this function, or a
better place to put it yet...