Better tracking of possessed entities and Zombies
Players now track their `possessed_entity`, a reference to
the full entity of the monster they took over.

This is now used to create more accurate Zombies of
monsters, with the same name and character.

(desaturated green is now exclusively used for Zombies,
with normal monsters getting their own shades of green
or brown.)

The Rendering function now displays the name of your
current body, centered over the Health Bar.